@@ -18,34 +18,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.0",
-    "@angular/common": "^4.3.0",
-    "@angular/compiler": "^4.3.0",
-    "@angular/core": "^4.3.0",
-    "@angular/forms": "^4.3.0",
-    "@angular/http": "^4.3.0",
-    "@angular/platform-browser": "^4.3.0",
-    "@angular/platform-browser-dynamic": "^4.3.0",
-    "@angular/platform-server": "^4.3.0",
-    "@angular/router": "^4.3.0",
-    "@angular/service-worker": "^1.0.0-beta.15",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^6.1.10",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
+    "@angular/platform-browser": "^6.1.10",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/platform-server": "^6.1.10",
+    "@angular/router": "^6.1.10",
+    "@angular/service-worker": "^6.1.10",
+    "core-js": "^2.5.7",
     "ng-pwa-tools": "0.0.15",
+    "node-fetch": "^2.2.0",
     "normalize.css": "^7.0.0",
-    "rxjs": "^5.1.0",
-    "unfetch": "^3.0.0",
-    "zone.js": "^0.8.13"
+    "rxjs": "^6.3.0",
+    "rxjs-compat": "^6.3.3",
+    "unfetch": "^4.0.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.2.0",
-    "@angular/compiler-cli": "^4.3.0",
-    "@angular/language-service": "^4.3.0",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "^6.2.5",
+    "@angular/compiler-cli": "^6.1.10",
+    "@angular/language-service": "^6.1.10",
     "@types/jasmine": "2.5.45",
     "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
+    "codelyzer": "~4.5.0",
     "gulp": "^3.9.1",
     "gulp-clean": "^0.3.2",
-    "gulp-debug": "^3.1.0",
+    "gulp-debug": "^4.0.0",
     "gulp-html-beautify": "^1.0.1",
     "gulp-string-replace": "^0.4.0",
     "gulp-useref": "2.1.0",
@@ -60,8 +63,11 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
     "run-sequence": "^2.0.0",
+    "rxjs-tslint": "^0.1.5",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "^2.7.2",
+    "webpack": "^4.21.0",
+    "webpack-dev-server": "^3.1.9"
   }
 }
